Move FocalPoint enum to file-utils
