+++ packages/api/cms-api/src/dam/blocks/pixel-image.block.ts
@@ -12,7 +12,7 @@ import {
     createBlock,
 } from "../../blocks/block";
 import { AnnotationBlockMeta } from "../../blocks/decorators/field";
-import { FocalPoint } from "../common/enums/focal-point.enum";
+import { FocalPoint } from "../../file-utils/focal-point.enum";
 import { FILE_ENTITY } from "../files/entities/file.entity";
 import { FilesService } from "../files/files.service";
 import { ImageCropAreaInput } from "../images/dto/image-crop-area.input";
