+++ packages/api/cms-api/src/file-utils/focal-point.enum.ts
